drivers: wifi: Combine all Wi-Fi management files
These are a single logical module implementing Zephyr's Wi-Fi management
operations. A file per-command is an overkill.
